Made login authentication state
Created login authentication state service to allow the login method to set the succesfully logged in user as the logged in user state. Allowing the UserPage to check whether an user is logged in or not and give access to information based on that.

Need to add Roles to the authentication service to allow for specific functions to be allowed or limited to different kind of users based on their roles.